backend/requirements.txt
django==4.0.1
djangorestframework==3.13.1
django-ckeditor==6.2.0
Pillow==9.0.0